Removal of durable storage of fingerprints
The complexity is drastically increased by another storage just to minimise a small amount of breakage in websites that may be using a fingerprint as a persistent identifier.

The remaining work would be to have accessor observers in the content script to verify the fingerprint has been used and message back to the background script, this would then updage the storage to ensure it's not cleared.

This additional code makes the storage lookup async which is a negative but also the aditional complexity would make a lot of edge cases to test.